Extract review queue transport helpers
